+++ addons/ammunition/cup/CfgMagazineWells.hpp
@@ -1,0 +1,223 @@
+class CfgMagazineWells {
+    // Handguns
+    class CBA_9x19_HiPower {
+        ADDON[] = {
+            QCLASS(13Rnd_9x19_Browning_Ball)
+        };
+    };
+    class CBA_9x19_CZ75_FULL {
+        ADDON[] = {
+            QCLASS(18Rnd_9x19_CZ75_Ball)
+        };
+    };
+    class CBA_9x19_Glock_Full {
+        ADDON[] = {
+            QCLASS(17Rnd_9x19_Glock_Ball)
+        };
+    };
+
+    class CBA_9x19_M17 {
+        ADDON[] = {
+            QCLASS(17Rnd_9x19_M17_Ball)
+        };
+    };
+
+    class CBA_9x19_M9 {
+        ADDON[] = {
+            QCLASS(15Rnd_9x19_M9_Ball)
+        };
+    };
+    class CBA_45ACP_MK23 {
+        ADDON[] = {
+            QCLASS(12Rnd_45ACP_MK23_Ball)
+        };
+    };
+
+    class CBA_45ACP_1911 {
+        ADDON[] = {
+            QCLASS(8Rnd_45ACP_1911_Ball)
+        };
+    };
+
+    // Rifles
+    class CBA_9x19_MP5 {
+        ADDON[] = {
+            QCLASS(30Rnd_9x19_Ball_MP5),
+            QCLASS(30Rnd_9x19_Ball_Tracer_MP5),
+            QCLASS(30Rnd_9x19_EPR_MP5),
+            QCLASS(30Rnd_9x19_AP_MP5)
+        };
+    };
+    class CBA_9x19_Vityaz {
+        ADDON[] = {
+            QCLASS(30Rnd_9x19_Ball_Vityaz),
+            QCLASS(30Rnd_9x19_Ball_Tracer_Vityaz),
+            QCLASS(30Rnd_9x19_EPR_Vityaz),
+            QCLASS(30Rnd_9x19_AP_Vityaz)
+        };
+    };
+    class CBA_9x19_PP19 {
+        ADDON[] = {
+            QCLASS(64Rnd_9x19_Ball_Bizon),
+            QCLASS(64Rnd_9x19_Ball_Tracer_Bizon),
+            QCLASS(64Rnd_9x19_EPR_Bizon),
+            QCLASS(64Rnd_9x19_AP_Bizon)
+        };
+    };
+    class CBA_12g_SAIGA {
+        ADDON[] = {
+            QCLASS(8Rnd_Saiga_000),
+            QCLASS(8Rnd_Saiga_AP20),
+            QCLASS(8Rnd_Saiga_Smoke),
+            QCLASS(20Rnd_Saiga_000),
+            QCLASS(20Rnd_Saiga_AP20),
+            QCLASS(20Rnd_Saiga_Smoke)
+        };
+    };
+    class CBA_46x30_MP7 {
+        ADDON[] = {
+            QCLASS(40Rnd_46x30_Ball),
+            QCLASS(40Rnd_46x30_Ball_Tracer),
+            QCLASS(40Rnd_46x30_EPR),
+            QCLASS(40Rnd_46x30_AP)
+        };
+    };
+    class CBA_68SPC_STANAG {
+        ADDON[] = {
+            QCLASS(30Rnd_68x43_Ball_ACR),
+            QCLASS(30Rnd_68x43_Ball_Tracer_ACR),
+            QCLASS(30Rnd_68x43_EPR_ACR),
+            QCLASS(30Rnd_68x43_AP_ACR)
+        };
+    };
+    class CBA_545x39_AK {
+        ADDON[] = {
+            QCLASS(60Rnd_545x39_Ball_Quadstack),
+            QCLASS(60Rnd_545x39_Ball_Tracer_Quadstack),
+            QCLASS(60Rnd_545x39_EPR_Quadstack)
+        };
+    };
+    class CBA_545x39_Fort {
+        ADDON[] = {
+            QCLASS(30Rnd_545x39_Ball_Fort),
+            QCLASS(30Rnd_545x39_Ball_Tracer_Fort),
+            QCLASS(30Rnd_545x39_EPR_Fort),
+            QCLASS(30Rnd_545x39_AP_Fort)
+        };
+    };
+    class CBA_556x45_STANAG {
+        ADDON[] = {
+            QCLASS(30Rnd_556x45_Ball_EMAG),
+            QCLASS(30Rnd_556x45_Ball_Tracer_EMAG),
+            QCLASS(30Rnd_556x45_EPR_EMAG),
+            QCLASS(30Rnd_556x45_AP_EMAG),
+            QCLASS(30Rnd_556x45_Ball_EMAG_Tan),
+            QCLASS(30Rnd_556x45_Ball_Tracer_EMAG_Tan),
+            QCLASS(30Rnd_556x45_EPR_EMAG_Tan),
+            QCLASS(30Rnd_556x45_AP_EMAG_Tan),
+            QCLASS(30Rnd_556x45_Ball_PMAG),
+            QCLASS(30Rnd_556x45_Ball_Tracer_PMAG),
+            QCLASS(30Rnd_556x45_EPR_PMAG),
+            QCLASS(30Rnd_556x45_AP_PMAG),
+            QCLASS(30Rnd_556x45_Ball_PMAG_Tan),
+            QCLASS(30Rnd_556x45_Ball_Tracer_PMAG_Tan),
+            QCLASS(30Rnd_556x45_EPR_PMAG_Tan),
+            QCLASS(30Rnd_556x45_AP_PMAG_Tan),
+            QCLASS(60Rnd_556x45_Ball_Surefire),
+            QCLASS(60Rnd_556x45_Ball_Tracer_Surefire),
+            QCLASS(60Rnd_556x45_EPR_Surefire),
+            QCLASS(100Rnd_556x45_Ball_BetaC),
+            QCLASS(100Rnd_556x45_Ball_Tracer_BetaC),
+            QCLASS(100Rnd_556x45_EPR_BetaC)
+        };
+    };
+    class CBA_556x45_AK {
+        ADDON[] = {
+            QCLASS(30Rnd_556x45_Ball_AK),
+            QCLASS(30Rnd_556x45_Ball_Tracer_AK),
+            QCLASS(30Rnd_556x45_EPR_AK),
+            QCLASS(30Rnd_556x45_AP_AK)
+        };
+    };
+    class CBA_556x45_G36 {
+        ADDON[] = {
+            QCLASS(30Rnd_556x45_Ball_G36),
+            QCLASS(30Rnd_556x45_Ball_Tracer_G36),
+            QCLASS(30Rnd_556x45_EPR_G36),
+            QCLASS(30Rnd_556x45_AP_G36),
+            QCLASS(100Rnd_556x45_Ball_BetaC),
+            QCLASS(100Rnd_556x45_Ball_Tracer_BetaC),
+            QCLASS(100Rnd_556x45_EPR_BetaC)
+        };
+    };
+
+    class CBA_556x45_FAMAS {
+        ADDON[] = {
+            QCLASS(25Rnd_556x45_Ball_FAMAS),
+            QCLASS(25Rnd_556x45_Ball_Tracer_FAMAS),
+            QCLASS(25Rnd_556x45_EPR_FAMAS),
+            QCLASS(25Rnd_556x45_AP_FAMAS)
+        };
+    };
+    class CBA_556x45_MINIMI {
+        ADDON[] = {
+            QCLASS(100Rnd_556x54_Ball_Pouch),
+            QCLASS(100Rnd_556x54_Ball_Tracer_Pouch),
+            QCLASS(100Rnd_556x54_EPR_Pouch),
+            QCLASS(200Rnd_556x45_Ball_Pouch),
+            QCLASS(200Rnd_556x45_Ball_Tracer_Pouch),
+            QCLASS(200Rnd_556x45_EPR_Pouch),
+            QCLASS(200Rnd_556x45_Ball_Box),
+            QCLASS(200Rnd_556x45_Ball_Tracer_Box),
+            QCLASS(200Rnd_556x45_EPR_Box)
+        };
+    };
+    class CLASS(762x39_CZ) {
+        ADDON[] = {
+            QCLASS(30Rnd_762x39_Ball_CZ),
+            QCLASS(30Rnd_762x39_Ball_Tracer_CZ),
+            QCLASS(30Rnd_762x39_EPR_CZ),
+            QCLASS(30Rnd_762x39_AP_CZ)
+        };
+    };
+    class CBA_762x51_FAL {
+        ADDON[] = {
+            QCLASS(20Rnd_762x51_Ball_FAL),
+            QCLASS(20Rnd_762x51_Ball_Tracer_FAL),
+            QCLASS(20Rnd_762x51_EPR_FAL),
+            QCLASS(20Rnd_762x51_AP_FAL)
+        };
+    };
+    class CBA_762x51_HK417 {
+        ADDON[] = {
+            QCLASS(20Rnd_762x51_Ball_HK417),
+            QCLASS(20Rnd_762x51_Ball_Tracer_HK417),
+            QCLASS(20Rnd_762x51_EPR_HK417),
+            QCLASS(20Rnd_762x51_AP_HK417)
+        };
+    };
+    class CBA_762x51_SCAR {
+        ADDON[] = {
+            QCLASS(20Rnd_762x51_Ball_SCAR),
+            QCLASS(20Rnd_762x51_Ball_Tracer_SCAR),
+            QCLASS(20Rnd_762x51_EPR_SCAR),
+            QCLASS(20Rnd_762x51_AP_SCAR)
+        };
+    };
+    class CBA_762x51_SCAR_EGLM {
+        ADDON[] = {
+            QCLASS(20Rnd_762x51_Ball_SCAR),
+            QCLASS(20Rnd_762x51_Ball_Tracer_SCAR),
+            QCLASS(20Rnd_762x51_EPR_SCAR),
+            QCLASS(20Rnd_762x51_AP_SCAR)
+        };
+    };
+    class CBA_9x39_VSS {
+        ADDON[] = {
+            QCLASS(30Rnd_9x39_Ball),
+            QCLASS(30Rnd_9x39_Ball_Tracer),
+            QCLASS(30Rnd_9x39_EPR),
+            QCLASS(30Rnd_9x39_AP)
+        };
+    };
+};
