+++ addons/ammunition/cup/config.cpp
@@ -21,3 +21,4 @@ class CfgPatches {
 
 #include "CfgAmmo.hpp"
 #include "CfgMagazines.hpp"
+#include "CfgMagazineWells.hpp"
