+++ addons/ammunition/ws/CfgMagazineWells.hpp
@@ -1,0 +1,38 @@
+class CfgMagazineWells {
+    class 556x45_Velko {
+        ADDON[] = {
+            QCLASS(35Rnd_556x45_Ball_R4),
+            QCLASS(35Rnd_556x45_Ball_Tracer_R4),
+            QCLASS(35Rnd_556x45_EPR_R4),
+            QCLASS(35Rnd_556x45_AP_R4),
+            QCLASS(50Rnd_556x45_Ball_R4),
+            QCLASS(50Rnd_556x45_Ball_Tracer_R4),
+            QCLASS(50Rnd_556x45_EPR_R4)
+        };
+    };
+    class AA40_12Gauge {
+        ADDON[] = {
+            QCLASS(8Rnd_AA40_000),
+            QCLASS(8Rnd_AA40_AP20),
+            QCLASS(8Rnd_AA40_Smoke),
+            QCLASS(20Rnd_AA40_000),
+            QCLASS(20Rnd_AA40_AP20),
+            QCLASS(20Rnd_AA40_Smoke)
+        };
+    };
+    class SLR_762x51 {
+        ADDON[] = {
+            QCLASS(20Rnd_762x51_Ball_SLR),
+            QCLASS(20Rnd_762x51_Ball_Tracer_SLR),
+            QCLASS(20Rnd_762x51_EPR_SLR),
+            QCLASS(20Rnd_762x51_AP_SLR)
+        };
+    };
+    class CBA_762x51_LINKS {
+        ADDON[] = {
+            QCLASS(100Rnd_762x51_Ball_Belt_SA77),
+            QCLASS(100Rnd_762x51_Ball_Tracer_Belt_SA77),
+            QCLASS(100Rnd_762x51_EPR_Belt_SA77)
+        };
+    };
+};
