+++ addons/weapons/CfgMagazineWells.hpp
@@ -4,4 +4,15 @@ class CfgMagazineWells {
             QCLASS(20Rnd_762x51_RFB_Mag)
         };
     };
+    class CLASS(Cylinder_357) {
+        ADDON[] = {
+            QCLASS(6Rnd_357)
+        };
+    };
+
+    class CLASS(Cylinder_44) {
+        ADDON[] = {
+            QCLASS(6Rnd_44)
+        };
+    };
 };
